pyproject.toml
[project]
name = "python-scaffold-template"
version = "0.1.0"
description = "Python 项目脚手架模板，提供基础的工程化目录布局。"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "psycopg2-binary>=2.9.10",
]
